feat(pattern): name change and feature-flag removed
