+++ packages/react/src/internal/FeatureFlags.js
@@ -80,14 +80,6 @@ export const DDS_SIMPLEBENEFITS =
 export const DDS_SIMPLELONGFORM =
   process.env.DDS_SIMPLELONGFORM === 'true' || DDS_FLAGS_ALL || false;
 
-/**
- * This determines if the simplelongform will be rendered or not
- *
- * @type {string | boolean}
- */
-export const DDS_CARD_ARRAY =
-  process.env.DDS_CARD_ARRAY === 'true' || DDS_FLAGS_ALL || false;
-
 /** This determines if the contentarraywithpictograms will be rendered or not
  *
  * @type {string | boolean}
@@ -133,14 +125,6 @@ export const DDS_LOGO_GRID =
  */
 export const DDS_FEATURED_LINK =
   process.env.DDS_FEATURED_LINK === 'true' || DDS_FLAGS_ALL || false;
-
-/**
- * This determines if the ContentGroup component will be rendered or not
- *
- * @type {string | boolean}
- */
-export const DDS_CONTENT_GROUP =
-  process.env.DDS_CONTENT_GROUP === 'true' || DDS_FLAGS_ALL || false;
 
 /**
  * This flag turns on/off the ButtonGroup component
