--- packages/react/src/patterns/blocks/CardArray/__stories__/index.scss
@@ -1,1 +1,0 @@
-@import '../../../../../../styles/scss/patterns/blocks/cardarray/index';
