+++ packages/react/src/patterns/blocks/ContentGroupCards/__stories__/index.scss
@@ -1,0 +1,6 @@
+@import '../../../../../../styles/scss/patterns/blocks/contentgroupcards/index';
+
+.#{$prefix}--contentgroupcards .#{$prefix}--content-group__children {
+  height: auto;
+  background-color: none;
+}
