+++ packages/react/src/patterns/blocks/ContentGroupCards/index.js
@@ -5,4 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-export { default as CardArray } from './CardArray';
+export { default as ContentGroupCards } from './ContentGroupCards';
