+++ packages/styles/scss/ibm-dotcom-styles.scss
@@ -27,7 +27,7 @@
 @import 'patterns/sections/simpleoverview/simpleoverview';
 
 // Patterns - Blocks
-@import 'patterns/blocks/cardarray/cardarray';
+@import 'patterns/blocks/contentgroupcards/contentgroupcards';
 @import 'patterns/blocks/FeaturedLink/featuredlink';
 @import 'patterns/blocks/listsection/listsection';
 @import 'patterns/blocks/logo-grid/logo-grid';
