--- packages/styles/scss/patterns/blocks/cardarray/README.md
@@ -1,32 +1,0 @@
-### SCSS
-
-#### Usage
-
-Import the package css into the top of your main CSS file.
-
-```css
-@import '@carbon/ibmdotcom-styles/scss/patterns/blocks/cardarray/index';
-```
-
-#### Elements
-
-| Class                               | Description                                                                            |
-| ----------------------------------- | -------------------------------------------------------------------------------------- |
-| `.bx--cardarray__container`         | Element containing all list section content                                            |
-| `.bx--cardarray__row`               | Vertical divisions across the viewport                                                 |
-| `.bx--cardarray__col`               | Horizonal divisions across the viewport and change in width at different screen widths |
-| `.bx--cardarray__title`             | Title for list section pattern                                                         |
-| `.bx--cardarray__divider_col`       | Horizonal division for divider element                                                 |
-| `.bx--cardarray__divider`           | Divider element                                                                        |
-| `.bx--cardarray-item__title`        | list section item title element                                                        |
-| `.bx--cardarray-item__content`      | list section item content element                                                      |
-| `.bx--cardarray-item__link`         | list section item Link element                                                         |
-| `.bx--cardarray-item__link__button` | list section item Link button element                                                  |
-
-#### Blocks
-
-| Class                  | Description                                                       |
-| ---------------------- | ----------------------------------------------------------------- |
-| `.bx--cardarray`       | Highest level class name for the Content Array with cards pattern |
-| `.bx--cardarray-group` | Content Array with cards section group component class            |
-| `.bx--cardarray-item`  | Content Array with cards section item component class             |
