+++ packages/styles/scss/patterns/blocks/contentgroupcards/README.md
@@ -1,0 +1,32 @@
+### SCSS
+
+#### Usage
+
+Import the package css into the top of your main CSS file.
+
+```css
+@import '@carbon/ibmdotcom-styles/scss/patterns/blocks/contentgroupcards/index';
+```
+
+#### Elements
+
+| Class                                       | Description                                                                            |
+| ------------------------------------------- | -------------------------------------------------------------------------------------- |
+| `.bx--contentgroupcards__container`         | Element containing all list section content                                            |
+| `.bx--contentgroupcards__row`               | Vertical divisions across the viewport                                                 |
+| `.bx--contentgroupcards__col`               | Horizonal divisions across the viewport and change in width at different screen widths |
+| `.bx--contentgroupcards__title`             | Title for list section pattern                                                         |
+| `.bx--contentgroupcards__divider_col`       | Horizonal division for divider element                                                 |
+| `.bx--contentgroupcards__divider`           | Divider element                                                                        |
+| `.bx--contentgroupcards-item__title`        | list section item title element                                                        |
+| `.bx--contentgroupcards-item__content`      | list section item content element                                                      |
+| `.bx--contentgroupcards-item__link`         | list section item Link element                                                         |
+| `.bx--contentgroupcards-item__link__button` | list section item Link button element                                                  |
+
+#### Blocks
+
+| Class                          | Description                                                       |
+| ------------------------------ | ----------------------------------------------------------------- |
+| `.bx--contentgroupcards`       | Highest level class name for the Content Array with cards pattern |
+| `.bx--contentgroupcards-group` | Content Array with cards section group component class            |
+| `.bx--contentgroupcards-item`  | Content Array with cards section item component class             |
