+++ packages/styles/scss/patterns/blocks/contentgroupcards/index.scss
@@ -6,7 +6,7 @@
 //
 
 @import '../../../globals/imports';
-@import 'cardarray';
-@import 'cardarray-item';
+@import 'contentgroupcards';
+@import 'contentgroupcards-item';
 @import '../../sub-patterns/content-group/content-group';
 @import '@carbon/grid/scss/grid';
